+++ src/extension.ts
@@ -8,12 +8,7 @@ export function activate(context: vscode.ExtensionContext) {
         },
     ];
     context.subscriptions.push(
-        vscode.languages.registerCompletionItemProvider(
-            selector,
-            new PathAutocomplete(),
-            '/',
-            '\\',
-        ),
+        vscode.languages.registerCompletionItemProvider(selector, new PathAutocomplete(), '/', '\\'),
     );
 }